+++ 進級課題/SceneResults.cpp
@@ -71,7 +71,7 @@ void SceneResults::Update()
 
 		if (m_step >= m_timings.size())
 		{
-			m_step = m_timings.size(); 
+			m_step = static_cast<int>(m_timings.size()); 
 		}
 
 	}
